pyproject.toml
[project]
name = "legaltextdecoder"
version = "0.1.0"
description = "description"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]
